First prototype for Whack A Mole
If you want to test it, please open the Whack A Mole scene, press play.

Then click Game Manager on the inspector > Right click and select Start Whacka on the context menu.

For now you can only watch the game play itself - the code for hammer hit and what the mole should do is there already. (For now it instantly returns to underground, we have to talk what we wanna do later (polishing)